+++ acceptance/exec/exec_tool_common_params_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	dircopy "github.com/otiai10/copy"
-	"github.com/puppetlabs/pdkgo/acceptance/testutils"
+	"github.com/puppetlabs/pct/acceptance/testutils"
 	"github.com/stretchr/testify/assert"
 )