+++ acceptance/explain/explain_test.go
@@ -3,7 +3,7 @@ package explain_test
 import (
 	"testing"
 
-	"github.com/puppetlabs/pdkgo/acceptance/testutils"
+	"github.com/puppetlabs/pct/acceptance/testutils"
 	"github.com/stretchr/testify/assert"
 )